first commit
Basic layout, a View containing two Views. - no styling made yet.
First view contains Text Input and a Button contained in a View with Flexbox on.
Second view contains text only for now, but later on will be a list of goals that will be added using the elements of the first view.